requirements.txt
getpass
sqlite3
ruamel.yaml
yaml
datetime
gi
